repalce CollateBase to core.py
